+++ CHANGES.rst
@@ -5,6 +5,8 @@ Revision 0.1.3, XX-05-2017
 * INET-ADDRESS-MIB configured as pre-built at pysnmp codegen
 * JSON codegen produces "nodetype" element for OBJECT-TYPE
 * Fix to mibdump.py --destination-directory option
+* Fix to pysnmp and JSON code generators to properly refer to MIB module
+  defining particular MIB object
 
 Revision 0.1.2, 12-04-2017
 --------------------------